// kodikosERP-master/src/app/models/formatki/kluczerejestrform.model.ts
export interface KluczeRejestrform {
  menuKluczeUprawnienia?: boolean;
  menuDodajUprawnienia?: boolean;
  dodajUprSaAktywne?: boolean;
  menuSpis?: boolean;
  menuDodajSpis?: boolean;
  menuWydania?: boolean;
  menuDodajWydanie?: boolean;
  rejestracja?: boolean;
  comments?: boolean;
  status?: boolean;
  buttonDodajKontr?: boolean;
  buttonOfertaKontr?: boolean;
  buttonEdytujKontr?: boolean;
  buttonUsunKontr?: boolean;
  widziWszystkie?: boolean;
}